feat(analyze): Adjusts Analysis Manager to maybe notify new feature
